mengedit point.c

+++ point.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include "point.h"
 
-POINT MakePOINT (float X, float Y){
+POINT MakePOINT (int X, int Y){
     POINT pp;
     pp.X = X;
     pp.Y = Y;
@@ -10,13 +10,13 @@ POINT MakePOINT (float X, float Y){
 }
 
 void BacaPOINT (POINT * P) {
-    float x,y;
-    scanf("%f %f", &x, &y);
+    int x,y;
+    scanf("%d %d", &x, &y);
     *P = MakePOINT(x,y);
 }
 
 void TulisPOINT (POINT P){
-    printf("(%.2f,%.2f)", P.X, P.Y);
+    printf("(%d,%d)", P.X, P.Y);
 }
 
 boolean EQ (POINT p1, POINT p2){
